Added verification
Removes skins of rarity 0 if no image are found. This change would be to prevent "Sets" from getting thrown in our tables.

This could be problematic if they release more characters since Default Skins are of this rarity, but considering new characters doesnt seem like a foreseeable thing for the next season, I don't think it'll cause problems.

Also Bwiki is on the grind whenever a new character comes out, so its okay.